requirements.txt
attrs==22.2.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
cryptography==38.0.4
distlib==0.3.6
dropbox==12.0.2
duolingo-api==0.5.4
filelock==3.8.2
google-api-core==2.10.2
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.57.0
httplib2==0.20.4
idna==3.4
iniconfig==1.1.1
jsonschema==4.1.2
#logme==0.1.0
numpy==1.24.0
oauthlib==3.2.0
packaging==22.0
pandas==2.2.3
pip==23.3.2
pipenv==2022.12.19
platformdirs==2.6.0
pluggy==0.13.1
protobuf==4.21.12
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.6.3
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.0.19
toml==0.10.2
typer==0.3.2
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
virtualenv-clone==0.5.7
wheel==0.38.4
selenium==4.27.1
instaloader
